add e2e.py test for the test_score_service.
docker-compose.yml:
changes the port of the flask service to 5001 due to port conflict.

Dockerfile:
add copy of the Utils.py so the MainScores.py will have access to Scores.txt file name.

MainScores.py:
add "host='0.0.0.0'" to expose the app so it will be accessible from the web browser.

requirements.txt:
add selenium and webdriver_manager for the e2e test.